requirements.txt
colorama==0.4.6
loguru==0.7.3
mysql-connector-python==9.2.0
pyfiglet==1.0.2
